use a build tag to enable logs
